// site/@cliotech/fx-core-api/src/facade/enums/IWebEInstrumentMap.ts
export enum IWebEInstrumentMap {
  id = 0,
  amountGroupId = 1,
  factor = 2,
  hasSignals = 3,
  ccyPair = 4,
  tradable = 5,
  defaultDealSize = 6,
  signalName = 7,
  maxDeal = 8,
  SLMinDistance = 9,
  TPMinDistance = 10,
  DecimalDigit = 11,
  PipDigit = 12,
  SpecialFontStart = 13,
  SpecialFontLength = 14,
  assetTypeId = 15,
  expirationDate = 16,
  instrumentTypeId = 17,
  baseSymbolId = 18,
  otherSymbolId = 19,
  exchangeInstrumentName = 20,
  exchange = 21,
  contractMonthAndYear = 22,
  instrumentEnglishName = 23,
  weightedVolumeFactor = 24,
  eventDate = 25,
  eventAmount = 26,
  marketPriceTolerance = 27,
}